NGEN: draw popup boxes on screen in one step
On slower machines the buildup process of these gui elements can be
observed. To speed it up, the box is now rendered completely
in memory and then copied to the screen in one step.
This feature needs 32kb more memory which is critical on DOS machines
or DOSBox with a low cycle value.
Since this is usually often used, it needs to be fast.

Next step is rendering the strings on top of the box,
to copy box + text in one step.