SCHICK: use gs_travel_map_ptr globvar
This pointer was moved out of the game state for
the future development.

If fact this is just an alias for g_renderbuf_ptr
and is updated with every call of load_map().

There is also an implementation for EMS Memory (>= 1MB),
which stores the travel map there, but it was disabled for some reason.