SCHICK: use g_enemy_sheets[] globvar completely
This last commit on this variable had and effect on the
binaryaequivalencycodedifference value,
which went from 3813 up to 3929. :-(

Nevertheless the functionality of the DOS code seems unchanged,
since the differences occur in the Overlay-Management code only.
On modern platforms this doesn't matter at all.

In the BCC-DOS domain this will hopefully clearify itsself or remains
as a task which will be considered at a later point in time.