SCHICK: rework seg48:status_menu (BAE OVR 319 => 245)
The BCC may have used some optimizations to merge identical code in
switch statements. Sadly I wasn't able to figure out the correct
compiler options, so 3 portions of code had to be rewritten for the BCC
with inline asm.
The code here is now generated identical, of course with identical
behaviour.